Complete refactor of folder structure
