+++ src/client/java/com/armorhud/armorHud.java
@@ -1,8 +1,8 @@
 package com.armorhud;
 
+import com.armorhud.config.config;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
-import net.minecraft.text.Text;
 
 public class armorHud implements ClientModInitializer {
