+++ src/client/java/com/armorhud/util/armorHudModMenu.java
@@ -1,5 +1,6 @@
-package com.armorhud;
+package com.armorhud.util;
 
+import com.armorhud.config.configScreen;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
