+++ src/client/java/com/armorhud/util/armorHudRegistries.java
@@ -1,0 +1,4 @@
+package com.armorhud.util;
+
+public class armorHudRegistries {
+}
